Remove classic res
