+++ Res/Views/TeenageEng/VoiceTypeAndToneSettingsTeView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct VoiceTypeAndToneSettingsTeView: View {
     var dismissAction: () -> Void
     @Binding var activeModal: MainViewTeenageEng.ActiveModal?
-    @Binding var selectedOption: Option?
+    @Binding var selectedOption: OptionTe?
     @Binding var isModalStepTwoEnabled: Bool
     @ObservedObject var callManager: CallManager
     @ObservedObject var keyboardResponder: KeyboardResponder
